UPDATE: Adds bindings to update timezone and locale
## Update "generate locale" utility

Adds French and Italian. Also updates the way we produce the cirrencies
Still a very primitive tool.

## Minor fixes to UI and proper support for locales/timezones

Semicolon works as a separator and minimal fixes so that it works.
Proper reworking will be done at a later stage

## Adds "display language" setting to core

Formulas are now in different language.


To create a new model you need to specify the locale, timezone and display language
This is a bit annoying because you see things like:

```
let model = Model::new_empty("Workbook1", "en", "UTC", "en")?;
```
The two "en" mean different things the first is the Locale and the second one is the language.


The display language impacts formulas and it is a property of the user.
Two different users can look at the same workbook in two different
languages.

This is in contrast with Excel but similar to Google sheets
Note that `FORMULATEXT(ref)` will always return the formula en English.
Also the arguments of `INFO` and `CELL` only work in English.
For instance `ZELLE("color", C1)` will return the color in the German language not `ZELLE("farbe", C1)`

## Minor fix: use proper uppercase/lowercase